requirements.txt
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0
openai==1.51.0
python-docx==1.1.2
pandas==2.2.3
